attempting to run multiple tests again

+++ build/clean_iap.sh
@@ -3,4 +3,10 @@
 echo "Hello world"
 echo "UNITY ${UNITY_VERSION}"
 
-rm -rf ./client/Assets/Plugins/UnityPurchasing
+if [ "$UNITY_VERSION" != "2018.4.18f1" ]
+then
+    echo "Deleting Unity IAP Plugin folder"
+    rm -rf ./client/Assets/Plugins/UnityPurchasing
+else
+    echo "Leaving Unity IAP Plugin alone"
+fi
